[move-ide] Refactored auto-completions implementation to multiple files
